+++ CHANGELOG.md
@@ -25,7 +25,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -
 
 -->
-## [v2.3.13](https://github.com/TUDelftGeodesy/caroline/tree/main) (01-Jul-2025, [diff](https://github.com/TUDelftGeodesy/caroline/compare/b58137efb9ad9e7a497372fe7d8cb3e0d20cda46...main))
+## [v2.4.0](https://github.com/TUDelftGeodesy/caroline/tree/main) (02-Jul-2025, [diff](https://github.com/TUDelftGeodesy/caroline/compare/b58137efb9ad9e7a497372fe7d8cb3e0d20cda46...main))
+
+### Added:
+- `stm_generation` job
+
+### Fixed:
+- `crop_to_zarr` now generates in the correct directory instead of in `crop_to_raw`'s directories.
+- `crop_to_zarr`'s documentation no longer refers to Re-SLC
+
+
+## [v2.3.13](https://github.com/TUDelftGeodesy/caroline/tree/main) (02-Jul-2025, [diff](https://github.com/TUDelftGeodesy/caroline/compare/b58137efb9ad9e7a497372fe7d8cb3e0d20cda46...main))
 
 ### Changed:
 - `crop_to_zarr` now runs on 8 cores instead of 4.